requirements.txt
gradio
openai
requests
ai4bharat-transliteration